requirements.txt
pytest
pytest-html
pytest-dist
pytest-ordering
pytest-rerunfailures
pytest-base-url
allure-pytest